requirements.txt
langchain_google_genai
streamlit
langchain
langchain-community
python-dotenv
pymupdf
beautifulsoup4
requests